fix(params): avoid int overflow when computing retarget timespan
The block timespan (previousHeader - firstBlockOfPeriod time) was cast to
int before clamping. Since block times are uint32, the difference can
exceed Integer.MAX_VALUE and overflow, corrupting the timespan and the
computed difficulty target. Do the subtraction and clamp in long, casting
to int only after the value is clamped into the allowed range.

+++ core/src/main/java/org/bitcoinj/params/TestNet4Params.java
@@ -187,7 +187,7 @@ public class TestNet4Params extends AbstractBitcoinNetParams {
 
         Block firstBlockOfPeriod = findFirstBlockOfPeriod(previousHeader, blockStore);
         int periodTimespanSeconds =
-            clampToAllowedTimespan((int) (previousHeader.getTimeSeconds() - firstBlockOfPeriod.getTimeSeconds()));
+            clampToAllowedTimespan(previousHeader.getTimeSeconds() - firstBlockOfPeriod.getTimeSeconds());
 
         BigInteger newTarget = Utils.decodeCompactBits(firstBlockOfPeriod.getDifficultyTarget())
             .multiply(BigInteger.valueOf(periodTimespanSeconds))
@@ -257,7 +257,7 @@ public class TestNet4Params extends AbstractBitcoinNetParams {
     }
 
     /** Clamps the measured timespan to [targetTimespan/4, targetTimespan*4], per the standard retarget rules. */
-    private int clampToAllowedTimespan(final int timespanSeconds) {
+    private int clampToAllowedTimespan(final long timespanSeconds) {
         int minimumTimespan = getTargetTimespan() / 4;
         int maximumTimespan = getTargetTimespan() * 4;
         if (timespanSeconds < minimumTimespan) {
@@ -266,7 +266,7 @@ public class TestNet4Params extends AbstractBitcoinNetParams {
         if (timespanSeconds > maximumTimespan) {
             return maximumTimespan;
         }
-        return timespanSeconds;
+        return (int) timespanSeconds;
     }
 
     /**
